move trait to kimchi
