+++ kimchi/src/circuits/polynomials/endosclmul.rs
@@ -4,8 +4,8 @@
 
 use std::marker::PhantomData;
 
+use crate::curve::KimchiCurve;
 use ark_ff::{FftField, Field, One, Zero};
-use commitment_dlog::srs::KimchiCurve;
 
 use crate::circuits::constraints::ConstraintSystem;
 use crate::circuits::gate::CircuitGate;
