+++ kimchi/src/prover.rs
@@ -28,6 +28,7 @@ use crate::{
         RecursionChallenge,
     },
     prover_index::ProverIndex,
+    curve::KimchiCurve,
 };
 use ark_ec::ProjectiveCurve;
 use ark_ff::{FftField, Field, One, PrimeField, UniformRand, Zero};
@@ -38,7 +39,6 @@ use ark_poly::{
 use array_init::array_init;
 use commitment_dlog::{
     commitment::{b_poly_coefficients, BlindedCommitment, CommitmentCurve, PolyComm},
-    srs::KimchiCurve,
 };
 use itertools::Itertools;
 use o1_utils::ExtendedDensePolynomial as _;
